Moved Cleaner implementation to infrastructure making it generic
