+++ ChustaSoft.GamersPlatformUtils.Abstractions/Repositories/IFileDeleteRepository.cs
@@ -4,8 +4,8 @@ using System.Threading.Tasks;
 
 namespace ChustaSoft.GamersPlatformUtils.Abstractions
 {
-    public interface ICleaner
+    public interface IFileDeleteRepository
     {
-        Task<CleanResult> CleanAsync(IEnumerable<FileInfo> paths);
+        Task<CleanResult> DeleteAsync(IEnumerable<FileInfo> paths);
     }
 }
